refactor: insertPlayer returns index, reuse color slots
fix toggleDray return type uint8_t

+++ player.h
@@ -5,6 +5,6 @@
 
 void initPlayer(float x, float y, uint32_t color);
 void updatePlayer(float dirX, float dirY);
-uint32_t toggleDraw();
+uint8_t toggleDraw();
 
 #endif
